fox(i32): no need for i64

+++ src/infrastructure/src/repositories/pg_video_repository.rs
@@ -26,7 +26,7 @@ struct VideoRecord {
     video_url: String,
     thumbnail_url: String,
     preview_url: String,
-    duration_milliseconds: i64,
+    duration_milliseconds: i32,
     view_count: i64,
     like_count: i64,
     dislike_count: i64,
